Add files via upload
The objective of the project is to 
1) Registration :-  
i) First You have to register yourself through Sign Up. 
ii) We will come to Sign In page. In case you don't know the password or forget password then you will directly come to ResetPass page where you can reset your password. That means your new password will update in your database.
2) Booking of Ticket:-
i) We have to fill all the information which required for booking. According to that data will be inserted into database.
3)Cancellation of Ticket:-
i) If you want to delete the ticket, then you have to choose a ticket which you want to delete. According to that the data will delete from database.